+++ src/cmd/init.rs
@@ -49,6 +49,10 @@ impl Cmd {
                     "
 function {name}() {{
     previous_cmd=\"$(fc -ln -1)\"
+    export FIXIT_FNS=\"$(
+        declare -F | cut -d' ' -f3
+        alias | cut -d'=' -f1
+    )\"
     fixed_cmd=\"$({executable} fix \"$previous_cmd\")\"
     if [ \"$fixed_cmd\" != \"\" ]; then
         eval \"$fixed_cmd\"
@@ -63,6 +67,11 @@ function {name}() {{
                     "
 function {name} -d \"Fix your previous command\"
     set -l previous_cmd \"$history[1]\"
+    set -lx FIXIT_FNS \"(
+        functions | cut -d' ' -f1
+        alias | cut -d' ' -f2
+    )\"
+    echo \"$FIXIT_FNS\"
     {executable} fix \"$previous_cmd\" | read -l fixed_cmd
     if [ \"$fixed_cmd\" != \"\" ]
         commandline \"$fixed_cmd\"
@@ -76,6 +85,10 @@ end
                 "
 function {name}() {{
     previous_cmd=\"$(fc -ln -1)\"
+    export FIXIT_FNS=\"$(
+        print -l ${{(ok)functions}}
+        alias | cut -d'=' -f1
+    )\"
     fixed_cmd=\"$({executable} fix \"$previous_cmd\")\"
     if [ \"$fixed_cmd\" != \"\" ]; then
         eval \"$fixed_cmd\"